Applied change as suggested by Filip Stachowiak.
From email:

Hi Martin, the apply_cluster method is checking if an entry starts with given
cluster name, but when doing so, it may match too early. For example if clusters
are named "a", "abc", "abcde", it may conclude that "abcddd" starts with "a",
instead of "abc". To fix it, it is just enough to add cdirs.sort(key=len,
reverse=True) in line 23 of cluster.py. If you think it's valid, just include
it. Best regards, Filip Stachowiak